refactored a LOT of functions to be reusable. I started a new App file so I didnt get lost in the code I had written before. Making all these reusable functions will make end game functions a breeze.
